support for concurrent adds to memtable
Summary:
This diff adds support for concurrent adds to the skiplist and
hash_skiplist memtable implementations.  Memory allocation is made
thread-safe by the addition of a spinlock for concurrent allocations.
Concurrent memtable writes are made via an additional method and don't
impose a performance overhead on the non-concurrent case, so parallelism
can be selected on a per-batch basis.

Parallel writes is not currently compatible with inplace updates, update
callbacks, or delete filtering.

TODO: this diff passes unit tests when concurrent writes are not
enabled, but is still in progress

Test Plan:
1. unit tests
2. TODO: low-level stress tests of concurrent access
3. TODO: new unit tests
4. TODO: verify no perf regressions when disabled
5. TODO: characterize perf effect when enabled

Reviewers:

Subscribers: dhruba, kradhakrishnan