Reduce block cache misses when memtable is flushed
Summary:
Under logdevice workload, there were increased block cache
misses when a memtable was flushed because we were rebuilding iterators
in that case since the version set changed. This was exacerbated in the
case of iterate_upper_bound, since file iterators which were over the
iterate_upper_bound would have been deleted and are now brought back as
part of the Rebuild, only to be deleted again. We now renew the iterators
and only build iterators for files which are added and delete file
iterators for files which are deleted.

Test Plan: DBTestTailingIterator.TailingIteratorTrimSeekToNext

Reviewers: sdong, anthony, IslamAbdelRahman, igor, tnovak

Subscribers: leveldb, dhruba, march

Differential Revision: https://reviews.facebook.net/D50463